refactor(security): extract posture snapshotting into internal/security/diff
Seven security_diff_* files move to internal/security/diff. Its only sibling
dependency is httpsec — IsHTMLResponse picks which responses contribute headers,
ParseSingleCookie normalizes Set-Cookie — which is exactly the coupling commit
58264122 created when it split cookie parsing out to be shared between the cookie
check and the snapshot builder. That is why httpsec had to be a leaf.

De-stuttered: SecurityDiffManager -> Manager, SecuritySnapshot -> Snapshot,
SecurityCookie -> Cookie, SecurityDiffResult -> Result, SecurityChange -> Change,
SecurityDiffSummary -> Summary, trackedSecurityHeaders -> trackedHeaders.
HandleDiffSecurity keeps its name — it names the MCP action it serves. No JSON
tag changes.

The cookie-flag and duration-format branch tests stranded in the two root
coverage files move here as helpers_test.go. Worth flagging: those
TestFormatDuration_* cases are the only coverage util.FormatDuration has
anywhere — internal/util has no test for it — so they are kept rather than
deleted as a wrapper test.

diff lands at 96.1% coverage on its own tests.
