+++ internal/security/diff/types.go
@@ -1,8 +1,16 @@
+// types.go — Snapshot, change and diff-result models plus the tracked header set.
 // Purpose: Defines security diff domain models and manager configuration.
 // Why: Keeps core diff data structures stable while behavior is split into focused files.
 // Docs: docs/features/feature/security-hardening/index.md
 
-package security
+// Package diff captures named snapshots of a site's security posture — headers,
+// cookie flags, per-endpoint auth and transport scheme — and reports what
+// regressed or improved between two of them.
+//
+// It shares the HTTP primitives in internal/security/httpsec with the scan
+// package (IsHTMLResponse decides which responses contribute headers,
+// ParseSingleCookie normalizes Set-Cookie) and depends on no other sibling.
+package diff
 
 import (
 	"sync"
@@ -13,7 +21,7 @@ import (
 // Types
 // ============================================
 
-// SecurityDiffManager stores and compares security posture snapshots.
+// Manager stores and compares security posture snapshots.
 //
 // Invariants:
 // - snapshots map and order slice are mutated only under mu.
@@ -22,45 +30,45 @@ import (
 //
 // Failure semantics:
 // - Invalid snapshot names/actions are rejected with explicit errors.
-type SecurityDiffManager struct {
+type Manager struct {
 	mu        sync.RWMutex
-	snapshots map[string]*SecuritySnapshot
+	snapshots map[string]*Snapshot
 	order     []string // insertion order for LRU eviction
 	maxSnaps  int
 	ttl       time.Duration
 }
 
-// SecuritySnapshot captures normalized security posture at one instant.
+// Snapshot captures normalized security posture at one instant.
 //
 // Invariants:
 // - Header/cookie/auth/transport maps are keyed by normalized origin or endpoint strings.
-type SecuritySnapshot struct {
+type Snapshot struct {
 	Name      string                       `json:"name"`
 	TakenAt   time.Time                    `json:"taken_at"`
 	Headers   map[string]map[string]string `json:"headers"`   // origin -> headerName -> value
-	Cookies   map[string][]SecurityCookie  `json:"cookies"`   // origin -> cookies
+	Cookies   map[string][]Cookie          `json:"cookies"`   // origin -> cookies
 	Auth      map[string]bool              `json:"auth"`      // endpoint (method+url) -> has_auth
 	Transport map[string]string            `json:"transport"` // origin -> "https" or "http"
 }
 
-// SecurityCookie records cookie attributes for comparison.
-type SecurityCookie struct {
+// Cookie records cookie attributes for comparison.
+type Cookie struct {
 	Name     string `json:"name"`
 	HttpOnly bool   `json:"httponly"`
 	Secure   bool   `json:"secure"`
 	SameSite string `json:"samesite"`
 }
 
-// SecurityDiffResult is the comparison response.
-type SecurityDiffResult struct {
-	Verdict      string              `json:"verdict"` // "regressed", "improved", "unchanged"
-	Regressions  []SecurityChange    `json:"regressions"`
-	Improvements []SecurityChange    `json:"improvements"`
-	Summary      SecurityDiffSummary `json:"summary"`
+// Result is the comparison response.
+type Result struct {
+	Verdict      string   `json:"verdict"` // "regressed", "improved", "unchanged"
+	Regressions  []Change `json:"regressions"`
+	Improvements []Change `json:"improvements"`
+	Summary      Summary  `json:"summary"`
 }
 
-// SecurityChange describes a single security posture change.
-type SecurityChange struct {
+// Change describes a single security posture change.
+type Change struct {
 	Category       string `json:"category"` // "headers", "cookies", "auth", "transport"
 	Severity       string `json:"severity"` // "critical", "high", "warning", "info"
 	Origin         string `json:"origin,omitempty"`
@@ -74,16 +82,16 @@ type SecurityChange struct {
 	Recommendation string `json:"recommendation"`
 }
 
-// SecurityDiffSummary provides aggregate change counts.
-type SecurityDiffSummary struct {
+// Summary provides aggregate change counts.
+type Summary struct {
 	TotalRegressions  int            `json:"total_regressions"`
 	TotalImprovements int            `json:"total_improvements"`
 	BySeverity        map[string]int `json:"by_severity"`
 	ByCategory        map[string]int `json:"by_category"`
 }
 
-// SecuritySnapshotListEntry is a summary for the list response.
-type SecuritySnapshotListEntry struct {
+// SnapshotListEntry is a summary for the list response.
+type SnapshotListEntry struct {
 	Name    string `json:"name"`
 	TakenAt string `json:"taken_at"`
 	Age     string `json:"age"`
@@ -94,7 +102,7 @@ type SecuritySnapshotListEntry struct {
 // Security headers tracked for diff comparison
 // ============================================
 
-var trackedSecurityHeaders = []string{
+var trackedHeaders = []string{
 	"Strict-Transport-Security",
 	"X-Content-Type-Options",
 	"X-Frame-Options",
@@ -116,13 +124,13 @@ var headerRemovedRecommendations = map[string]string{
 // Constructor
 // ============================================
 
-// NewSecurityDiffManager creates a bounded snapshot store.
+// NewManager creates a bounded snapshot store.
 //
 // Invariants:
 // - Defaults favor short-lived regression checks (maxSnaps=5, ttl=4h).
-func NewSecurityDiffManager() *SecurityDiffManager {
-	return &SecurityDiffManager{
-		snapshots: make(map[string]*SecuritySnapshot),
+func NewManager() *Manager {
+	return &Manager{
+		snapshots: make(map[string]*Snapshot),
 		order:     make([]string, 0),
 		maxSnaps:  5,
 		ttl:       4 * time.Hour,
